# openapi/v1/openapi.yaml
openapi: 3.1.1
info:
  version: 0.1.0
  title: Cozy Creator
  description: Cozy Creator OpenAPI specification
servers:
  - url: https://api.cozy.art/v1
    description: Production endpoint
  - url: https://api.cozy.dev/v1
    description: Development-branch endpoint
...
components:
  parameters:
...
    AcceptStandard:
      name: Accept
      in: header
      schema:
        type: string
        enum:
          - application/json
          - application/vnd.msgpack
        default: application/json
      required: false
      description: Format the client wants to receive (defaults to JSON)
...
  securitySchemes:
    APIKeyHeader:
      in: header
      type: apiKey
      name: X-API-Key
  schemas:
    JobStatus:
      type: string
      enum: [IN_QUEUE, IN_PROGRESS, COMPLETED, FAILED]
      description: Status of the job
    JobStatusResponse:
      type: object
      properties:
        id:
          type: string
          description: ID of the Job
        status:
          $ref: "#/components/schemas/JobStatus"
      required: [id, status]
...
  responses:
...
    Error401:
      description: Unauthorized - invalid or missing API key
    Error500:
      description: Internal server error
paths:
...
  /jobs/{id}/status:
    get:
      parameters:
        - $ref: '#/components/parameters/AcceptStandard'
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The job ID
      tags:
        - generation
      summary: Retrieve the current status of a job
      operationId: getJobStatus
      security:
        - APIKeyHeader: []
      responses:
        "200":
          description: Job status retrieved successfully
          content:
            application/vnd.msgpack:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          description: Job not found
        "401":
          $ref: "#/components/responses/Error401"
        "500":
          $ref: "#/components/responses/Error500"
...